[REFACTOR] ViewingAvailableDateTime 변환 클래스 도입
- 기존 DTO와 뷰잉 가능 시간 생성을 분리시킨다
