+++ src/main/java/org/hanihome/hanihomebe/property/web/dto/request/create/PropertyCreateRequestDTO.java
@@ -42,7 +42,6 @@ public sealed interface PropertyCreateRequestDTO permits
 
     LivingConditions livingConditions();
 
-//    Set<LocalDateTime> availableFrom();
     MoveInInfo moveInInfo();
 
     LocalDate meetingDateFrom();
