+++ src/main/java/org/hanihome/hanihomebe/property/web/dto/request/create/SharePropertyCreateRequestDTO.java
@@ -4,13 +4,12 @@ import lombok.extern.slf4j.Slf4j;
 import org.hanihome.hanihomebe.global.exception.CustomException;
 import org.hanihome.hanihomebe.global.response.domain.ServiceCode;
 import org.hanihome.hanihomebe.interest.region.Region;
-import org.hanihome.hanihomebe.property.application.TimeSlotValidator;
+import org.hanihome.hanihomebe.property.application.time.generator.ViewingAvailableDateTimeGenerator;
+import org.hanihome.hanihomebe.property.application.time.validator.TimeSlotValidator;
 import org.hanihome.hanihomebe.property.domain.enums.*;
 import org.hanihome.hanihomebe.property.domain.vo.*;
-import org.hanihome.hanihomebe.viewing.domain.ViewingTimeInterval;
 
 import java.time.LocalDate;
-import java.time.LocalTime;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -59,28 +58,6 @@ public record SharePropertyCreateRequestDTO(
         }
 
         // ViewingAvailableDateTime 변환
-        // TODO: 해당 변환을 별도의 클래스에 위임하고 서비스에서 호출하는 것이 나아보임
-        LocalDate tempDate = meetingDateFrom;
-        while (tempDate.isBefore(meetingDateTo) || tempDate.isEqual(meetingDateTo)) {
-            log.info("현재 DTO 생성중의 date:{}", tempDate.toString() );
-            LocalDate finalTempDate = tempDate;
-            List<ViewingAvailableDateTime> finalViewingAvailableDateTimes = viewingAvailableDateTimes;
-            timeSlots.forEach(timeSlot -> {
-                LocalTime timeFrom = timeSlot.getTimeFrom();
-                LocalTime timeTo = timeSlot.getTimeTo();
-                while(timeFrom.isBefore(timeTo)) {
-                    ViewingAvailableDateTime viewingAvailableDateTime = new ViewingAvailableDateTime(finalTempDate,
-                            timeFrom,
-                            false,
-                            ViewingTimeInterval.MINUTE30);
-                    finalViewingAvailableDateTimes.add(viewingAvailableDateTime);
-                    timeFrom = timeFrom.plusMinutes(30);
-                }
-            });
-            tempDate = tempDate.plusDays(1);
-        }
-        log.info("meetingDateFrom: {}, meetingDateTo: {}", meetingDateFrom, meetingDateTo);
-        log.info("viewingAvailableDateTimes: {}", viewingAvailableDateTimes.stream().map(viewingAvailableDateTime -> viewingAvailableDateTime.getTime()).toList());
-
+        viewingAvailableDateTimes = ViewingAvailableDateTimeGenerator.generate(meetingDateFrom, meetingDateTo, timeSlots);
     }
 }
